Fixed classes

+++ Pizzawinkel/src/com/nhlstenden/decoratorPattern/Champignons.java
@@ -24,7 +24,7 @@ public class Champignons extends PizzaDecorator
     }
 
     @Override
-    public int getCost()
+    public double getCost()
     {
         return 0;
     }
